html数据双向绑定

+++ question_bank/src/main/java/com/czy/question/model/enums/OptionTypeEnum.java
@@ -6,24 +6,25 @@ import com.czy.util.enums.IEnum;
  * @author chenzy
  * @date 2020-06-22
  * 选项类型-英文字母、阿拉伯数字、判断对错、文本、其他填写
+ * 选项类型-英文字母、阿拉伯数字、判断对错、文本、其他
  */
-public enum OptionTypeEnum implements IEnum<Byte> {
-    letter((byte)1,"英文字母")
-    ,integer((byte)2,"阿拉伯数字")
-    ,judge((byte)3,"判断对错")
-    ,textFill((byte)2,"文本")
-    ,other((byte)3,"其他")
+public enum OptionTypeEnum implements IEnum<Integer> {
+    letter(1,"英文字母")
+    ,integer(2,"阿拉伯数字")
+    ,judge(3,"判断对错")
+    ,textFill(2,"文本")
+    ,other(3,"其他")
     ,;
 
-    private byte id;
+    private Integer id;
     private String msg;
-    OptionTypeEnum(byte id, String msg) {
+    OptionTypeEnum(Integer id, String msg) {
         this.id = id;
         this.msg = msg;
     }
 
     @Override
-    public Byte getValue() {
+    public Integer getValue() {
         return id;
     }
 }
